Problem: memiavl/versiondb don't build with upstream sdk
Solution:
- keep it api compatible with both
