+++ versiondb/objstore.go
@@ -1,0 +1,11 @@
+//go:build objstore
+// +build objstore
+
+package versiondb
+
+import "cosmossdk.io/store/types"
+
+// GetObjKVStore implements `MultiStore` interface
+func (s *MultiStore) GetObjKVStore(storeKey types.StoreKey) types.ObjKVStore {
+	return s.GetStore(storeKey).(types.ObjKVStore)
+}
